[BOOK-67] refactor: infra - 패키지 이동
- entity를 infra 모듈로 이동함
