+++ infra/src/main/kotlin/org/yapp/config/external/redis/RedisConfig.kt
@@ -1,4 +1,4 @@
-package org.yapp.redis
+package org.yapp.config.external.redis
 
 import org.springframework.beans.factory.annotation.Value
 import org.springframework.context.annotation.Bean
