+++ infra/src/main/kotlin/org/yapp/config/internal/jpa/JpaConfig.kt
@@ -1,4 +1,4 @@
-package org.yapp.jpa
+package org.yapp.config.internal.jpa
 
 import org.springframework.boot.autoconfigure.domain.EntityScan
 import org.springframework.data.jpa.repository.config.EnableJpaRepositories
@@ -6,8 +6,6 @@ import org.springframework.transaction.annotation.EnableTransactionManagement
 import org.yapp.InfraBaseConfig
 
 @EnableTransactionManagement
-@EntityScan(basePackages = ["org.yapp.domain.user"])
-@EnableJpaRepositories(basePackages = ["org.yapp.jpa.repository"])
-class JpaConfig : InfraBaseConfig {
-
-}
+@EntityScan(basePackages = ["org.yapp"])
+@EnableJpaRepositories(basePackages = ["org.yapp"])
+class JpaConfig : InfraBaseConfig
