+++ infra/src/main/kotlin/org/yapp/redis/repository/RedisTokenRepository.kt
@@ -1,4 +1,4 @@
-package org.yapp.redis
+package org.yapp.redis.repository
 
 import org.springframework.data.redis.core.RedisTemplate
 import org.springframework.stereotype.Repository
